♻️ Refactor: 드롭다운 폴더 구조를 팀 컨벤션에 맞게 분리
