+++ src/components/common/dropdown/index.ts
@@ -1,0 +1,2 @@
+export { default as ListDropdown } from '@/components/common/dropdown/components/ListDropdown';
+export { default as SelectDropdown } from '@/components/common/dropdown/components/SelectDropdown';
